mkfs: link using libf2fs.la
With slibtool the build will fail with 'cannot find -lf2fs'.

To fix this it is better to use the canonical method of linking internal
dependencies which is to use the generated .la files where the libtool
implementation will know what to do with it.

Gentoo-Issue: https://bugs.gentoo.org/929070
Signed-off-by: orbea <orbea@riseup.net>
Reviewed-by: Chao Yu <chao@kernel.org>
Signed-off-by: Jaegeuk Kim <jaegeuk@kernel.org>